Move `PieceGetter` trait to the root of `subspace-farmer-components` since it is not plotting-specific
