+++ crates/subspace-farmer-components/benches/plotting.rs
@@ -8,11 +8,9 @@ use subspace_core_primitives::crypto::kzg;
 use subspace_core_primitives::crypto::kzg::Kzg;
 use subspace_core_primitives::{HistorySize, PublicKey, Record, RecordedHistorySegment};
 use subspace_erasure_coding::ErasureCoding;
-use subspace_farmer_components::plotting::{
-    plot_sector, PieceGetterRetryPolicy, PlotSectorOptions,
-};
+use subspace_farmer_components::plotting::{plot_sector, PlotSectorOptions};
 use subspace_farmer_components::sector::sector_size;
-use subspace_farmer_components::FarmerProtocolInfo;
+use subspace_farmer_components::{FarmerProtocolInfo, PieceGetterRetryPolicy};
 use subspace_proof_of_space::chia::ChiaTable;
 use subspace_proof_of_space::Table;
